# shapetracker.py: better view/shapetracker caching