Make most View and Shapetracker methods static. These are mainly pushing tuples around, this should help with caching.
View is now a NamedTuple.

Ultimately ShapeTracker could become a similar Tuple, and each op creates a new one.

Still needs a lot of work on the tests, I'd like to hear if you think it's worth the effort @rayanht @geohot.


`test_net_speed.py`
Master:
```
27.90 lazy.py
15.84 tensor.py
27.06 shapetracker.py
15.83 ops.py
20.01 symbolic.py
9.06 helpers.py
1.07 lib.py
116.76 total
```

This branch:
```
26.30 lazy.py
13.51 tensor.py
16.10 shapetracker.py
13.09 ops.py
1.78 symbolic.py
6.87 helpers.py
0.60 lib.py
78.24 total
```