@tinyb0t 2023-06-25
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/codegen/linearizer.py            11              11
tinygrad/lazy.py                           1               1
tinygrad/shape/shapetracker.py           167             120
------------------------------------------------------------
total                                    179             132
------------------------------------------------------------
lines added in the tinygrad folder: 47
```