@roelofvandijk 2023-06-26
master (52b7105f8798d88feed0197691391f7b9aa4b013)
```
using <class 'tinygrad.runtime.ops_gpu.CLCodegen'>
testing llama python run time
built model
assigned empty tensors, doing warmup
codegen        runtime (median):  211.05ms, runs:   291.58,  210.03,  207.10,  205.42,  213.85,  255.18,  208.70,  205.04,  212.07,  220.11
methodcache    runtime (median):  216.98ms, runs:   219.24,  268.04,  210.51,  204.78,  244.21,  227.40,  204.73,  243.30,  212.41,  214.72
profile        runtime (median): 1106.56ms, runs:  1291.20, 1067.84, 1067.99, 1124.23, 1129.57, 1105.92, 1086.90, 1107.20, 1192.22, 1091.59
```


this branch
```
using <class 'tinygrad.runtime.ops_gpu.CLCodegen'>
testing llama python run time
built model
assigned empty tensors, doing warmup
codegen        runtime (median):  209.61ms, runs:   273.33,  204.58,  208.70,  194.39,  238.57,  210.01,  206.48,  228.19,  273.12,  209.20
methodcache    runtime (median):  204.24ms, runs:   195.37,  206.58,  246.79,  245.70,  237.84,  198.21,  201.90,  192.73,  194.53,  231.76
profile        runtime (median): 1010.56ms, runs:  1119.41, 1079.45, 1007.60, 1039.65, 1010.28,  997.83, 1010.84, 1001.20, 1116.74, 1004.89
```